--- carla_labels.json
@@ -1,25 +1,0 @@
-{
-    "Unlabeled": 0,	
-    "Building": 1,
-    "Fence": 2,
-    "Other": 3,
-    "Pedestrian": 4,	
-    "Pole":	5,
-    "RoadLine": 6,
-    "Road":	7,
-    "SideWalk": 8,	
-    "Vegetation": 9,		
-    "Vehicles": 10,
-    "Wall": 11,
-    "TrafficSign": 12,
-    "Sky": 13,
-    "Ground": 14,	
-    "Bridge": 15,	
-    "RailTrack": 16,	
-    "GuardRail": 17,
-    "TrafficLight": 18,	
-    "Static": 19,
-    "Dynamic": 20,
-    "Water"	: 21,
-    "Terrain": 22
-}
